reorganized project directories and import statements
